+++ adapters/qt/canvas_item.h
@@ -7,11 +7,20 @@
 #include <QQuickPaintedItem>
 class CanvasItem : public QQuickPaintedItem {
     Q_OBJECT
+    Q_PROPERTY(bool hasRegion READ hasRegion NOTIFY regionChanged)
+    Q_PROPERTY(int selectionMedia READ selectionMedia WRITE setSelectionMedia NOTIFY regionChanged)
+    Q_PROPERTY(QString regionInfo READ regionInfo NOTIFY regionChanged)
     Q_PROPERTY(EditorController* editor READ editor WRITE setEditor NOTIFY editorChanged)
     Q_PROPERTY(double zoom READ zoom WRITE setZoom NOTIFY viewChanged)
     Q_PROPERTY(bool mirrored READ mirrored WRITE setMirrored NOTIFY viewChanged)
     Q_PROPERTY(double rotationAngle READ rotationAngle WRITE setRotationAngle NOTIFY viewChanged)
   public:
+    bool hasRegion() const { return region_.width > 0 && region_.height > 0; }
+    int selectionMedia() const { return int(selectionMedia_); }
+    void setSelectionMedia(int);
+    QString regionInfo() const;
+    Q_INVOKABLE void clearRegion();
+    Q_INVOKABLE void transformRegion(int action, int dx = 0, int dy = 0);
     explicit CanvasItem(QQuickItem* parent = nullptr);
     EditorController* editor() const { return editor_; }
     void setEditor(EditorController*);
@@ -28,6 +37,7 @@ class CanvasItem : public QQuickPaintedItem {
     Q_INVOKABLE void cancelGesture();
     Q_INVOKABLE void capture(QString path);
   signals:
+    void regionChanged();
     void editorChanged();
     void viewChanged();
 
@@ -41,6 +51,9 @@ class CanvasItem : public QQuickPaintedItem {
     bool eventFilter(QObject*, QEvent*) override;
 
   private:
+    opentoon::PixelRect region_;
+    opentoon::SelectionMedia selectionMedia_ = opentoon::SelectionMedia::Both;
+    bool movingRegion_ = false;
     QPointer<EditorController> editor_;
     QPointer<QQuickWindow> filteredWindow_;
     double zoom_ = 1, angle_ = 0;